requirements.txt
Django==1.6.1
PIL==1.1.7
Pygments==1.6
Werkzeug==0.9.4
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-extensions==1.2.5
django-pygments==0.1
django-redis-cache==0.10.1
django-toolbelt==0.0.1
djangorestframework==2.3.10
gunicorn==18.0
psycopg2==2.5.1
redis==2.8.0
six==1.4.1
static==0.4
wsgiref==0.1.2
yolk==0.4.3
